Restructure parser so that a single function is responsible for creating the normalizr schema
