+++ src/server/graphql/schema/resolvers/SessionResolvers.js
@@ -1,8 +1,8 @@
 'use strict';
 
 
-const Session = require('../../models/Session');
-const Activity = require('../../models/Activity');
+const Session = require('../../../models/Session');
+const Activity = require('../../../models/Activity');
 
 const sessionResolvers = {
   Query: {
