+++ src/server/graphql/schema/resolvers/UserResolvers.js
@@ -1,9 +1,8 @@
 'use strict';
 
-
-const User = require('../../models/User');
-const Session = require('../../models/Session');
-const Category = require('../../models/Category');
+const User = require('../../../models/User');
+const Session = require('../../../models/Session');
+const Category = require('../../../models/Category');
 
 const userResolvers = {
   Query: {
